Template and Script Handling Improvements

+++ roles/water_gkhs_filebot/templates/clear_bullshit.j2.sh
@@ -9,6 +9,11 @@ else
 	FILEPATH=$1
 fi
 
+if [ ! -d "$FILEPATH/completed" ]; then
+	echo "Directory not found at $FILEPATH/completed"
+	exit 5
+fi
+
 find "$FILEPATH/completed" -iname ".xattr" -type d -exec rm -rf {} \;
 find "$FILEPATH/completed" -links +1 -type f -delete -print
 find "$FILEPATH/completed" -iname "screen*.jpg" -type f -delete -print
@@ -20,5 +25,5 @@ find "$FILEPATH/completed" -iname "*srt" -delete -print
 find "$FILEPATH/completed" -iname "RARBG.txt" -delete -print
 find "$FILEPATH/completed" -iname "RARBG_DO_NOT_MIRROR.exe" -delete -print
 find "$FILEPATH/completed" -iname "*tvm.jpg" -delete -print
-find "$FILEPATH/completed" -type d -exec rmdir {} \;
+find "$FILEPATH/completed" -depth -type d -empty -delete
 find "$FILEPATH/completed" -iname "www.YTS.MX.jpg" -delete -print
